L3 layout: RecalcHash[Next] int* → byte* (58KB×2 → 14KB×2; +0.5% avg)
Same pattern as the `_inGroup byte*` win (commit e6824e5): 0/1 dedup
flags fit byte, halve the array footprint while keeping straight
load/store semantics (no shift+mask).

Bit-identical (checksum 0x933ABE7915AC18BE on full_palette.nes 50K hc).

A/B bench on top of NodeInfo SoA split:
  prior:           best 39,145  avg 38,811  hc/s
  byte RecalcHash: best 39,362  avg 39,011  hc/s
                   → +0.5% best, +0.5% avg

Tried ulong* bitset first (1.8 KB × 2) but the shift+mask per-access
cost erased the cache benefit — net flat-to-slightly-negative on avg.
byte* is the right intermediate: cache-hot already (28 KB total fits
L1d alongside NodeStates + _inGroup + the new byte hashes — ~56 KB
all-in for 6 dedup/state arrays), with zero per-access overhead.

Cumulative L3-layout series vs original aot-codegen baseline:
                          best       avg
  original:               37,928    36,923
  byte _inGroup:          38,534    38,018  (+3.0%)
  + ushort _groupBuf:     38,591    37,874  (+2.6%)
  + ushort TList:         39,056    38,587  (+4.5%)
  + NodeInfo SoA split:   39,145    38,811  (+5.1%)
  + byte RecalcHash:      39,362    39,011  (+5.7%)  ← here

Files: WireCore.cs (decl + alloc), WireCore.Recalc.cs (EnqueueNode,
ProcessQueueInterp swap + inner loop), WireCore.Levelize.cs (swap +
inner loop), WireCore.Native.cs (null assignment grouping). Group.cs
/ GroupSimd.cs / Oblivious.cs unchanged — byte* makes the original
`RecalcHash[nn] = 0` access pattern work identically.

Confirms the access-pattern rule once more: random-access byte arrays
win when straight load suffices; bitset wins only when memory pressure
is high enough to dominate the shift+mask cost — and here our working
set already fits L1d, so bitset has nothing to offer.

+++ src/AprVisual/Sim/WireCore.Levelize.cs
@@ -101,7 +101,7 @@ namespace AprVisual.Sim
 
                 // swap "next" ↔ "current"
                 int* tmpList = RecalcList; RecalcList = RecalcListNext; RecalcListNext = tmpList;
-                int* tmpHash = RecalcHash; RecalcHash = RecalcHashNext; RecalcHashNext = tmpHash;
+                byte* tmpHash = RecalcHash; RecalcHash = RecalcHashNext; RecalcHashNext = tmpHash;
                 RecalcListCount = RecalcListNextCount;
                 RecalcListNextCount = 0;
